requirements.txt
bitstring==3.1.6
ndspy==3.0.0
Pillow-SIMD==6.0.0.post0
